requirements.txt
fastapi
uvicorn[standard]
python-dotenv
psycopg2-binary
supabase